% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{definition}
\alias{definition}
\alias{define}
\alias{is.definition}
\alias{print.definition}
\title{Specify a population model and priors for inference on vital rates}
\usage{
define(
  x,
  masks = list(),
  priors = list(),
  stat = aae.pop.inference::stat_abundance_trend,
  ...
)

is.definition(x)

\method{print}{definition}(x, ...)
}
\arguments{
\item{x}{an object returned from
\code{\link[aae.pop:dynamics]{aae.pop::dynamics()}} containing
a stochastic matrix population model}

\item{masks}{a logical matrix or vector (or list of these) defining cells
to be estimated. See \code{\link[aae.pop:masks]{aae.pop::masks()}} for
details}

\item{priors}{a named list of prior distributions, with one element for
survival and another for reproduction. Each element can be a single
character vector (see Details) or a list of character vectors. The
former assumes all survival/reproduction priors are identical, while
the latter allows different priors for each parameter}

\item{stat}{a function to summarise the output of a population dynamics
model run with \code{\link[aae.pop:simulate]{aae.pop::simulate()}}.
Defaults to
\code{\link[aae.pop.inference:stat_abundance_trend]{aae.pop.inference::stat_abundance_trend()}}}

\item{\dots}{additional arguments passed to
\code{\link[aae.pop:simulate]{aae.pop::simulate()}}}
}
\description{
Return a simulation function and suitable priors to perform
  inference on the vital rates component of a stochastic matrix population
  model with the \code{aae.pop.inference} package.
}
\details{
\code{define} returns a pre-defined template that can be used with
  \code{\link[aae.pop.inference:inference]{aae.pop.inference::inference()}}
  to estimate the demographic parameters of a stochastic population model.
  More complex models will require a custom specification but \code{define}
  is a useful placeholder in simpler cases where model arguments are fixed
  and only the demographic vital rates are variable.

  Priors should be specified as described in
  \code{\link[EasyABC:ABC_sequential]{EasyABC::ABC_sequential()}}.
  Currently supported options are the uniform, normal, lognormal,
  and exponential distributions, and each list element should be a
  vector specifying the distribution ("unif", "normal", "lognormal",
  or "exponential"), followed by one (for the exponential distribution)
  or two parameters (all other distributions)
}
\examples{
# setup: define a pop model to simulate and use in ABC
nclass <- 5
popmat <- matrix(0, ncol = nclass, nrow = nclass)
popmat[transition(popmat)] <- c(0.3, 0.5, 0.7, 0.8)
popmat[reproduction(popmat)] <- c(0, 1, 4, 10)
dd <- density_dependence(
  funs = ricker(100, exclude = 1),
  masks = reproduction(popmat)
)
pop_fn <- dynamics(matrix = popmat, dd)

# simulate some data from this model to use as a target data set
obs <- simulate(
  pop_fn,
  nsim = 100,
  args = list(density_dependence = list(theta = 0.4))
)

# define the inference model for all non-zero parameters
#   (nsim too low for accurate results)
pop_inf <- define(
  x = pop_fn,
  masks = list(
    survival = aae.pop::transition(pop_fn$matrix),
    reproduction = aae.pop::reproduction(pop_fn$matrix)
  ),
  nsim = 5
)

# and run it (quick run, nb_simul too low and p_acc_min too high for
#   accurate results)
pars <- aae.pop.inference::inference(
  model = pop_inf$model,
  prior = pop_inf$prior,
  target = pop_inf$stat(obs),
  nb_simul = 100,
  progress_bar = TRUE,
  p_acc_min = 0.3
)

# plot estimated parameters (not expected to be accurate)
plot(pars)
}
