% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference}
\alias{inference}
\alias{is.inference}
\title{Estimate parameters of a simulation model with sequential
  approximate Bayesian computation}
\usage{
inference(model, prior, target, method = NULL, ...)

is.inference(x)
}
\arguments{
\item{model}{a function to simulate a population dynamics object
from a set of parameters. This function must return a
vector of summary statistics with the same length as \code{target}}

\item{prior}{a list of prior distributions for each parameter, specified
as described in
\code{\link[EasyABC:ABC_sequential]{EasyABC::ABC_sequential()}}.
Currently supported options are the uniform, normal, lognormal,
and exponential distributions}

\item{target}{values to be compared to model outputs simulated
with \code{model}}

\item{\dots}{additional arguments passed to the sequential ABC sampler
(see \code{\link[EasyABC:ABC_sequential]{EasyABC::ABC_sequential()}})}
}
\description{
Generic function to estimate parameters using sequential
  ABC, targeted specifically towards models implemented with the
  \code{\link{simulate}} function in the \code{aae.pop} package.
}
\details{
Approximate Bayesian computation is a fast method for parameter
  inference in cases where model likelihoods are intractable (or at
  least difficult to compute). The \code{inference} function is a wrapper
  for the \code{\link[EasyABC:ABC_sequential]{EasyABC::ABC_sequential()}}
  function intended to simplify parameter inference for population
  dynamics models created with
  \code{\link[aae.pop:dynamics]{aae.pop::dynamics()}}.

  Although standard MCMC algorithms are feasible for many of these models,
  the use of ABC allows fast and flexible specification of models based on
  existing simulation tools such as those in the \code{aae.pop} package.
}
\examples{
# to add
}
