% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{statistics}
\alias{statistics}
\alias{stat_abundance_trend}
\alias{stat_abundance_moment}
\title{Calculate summary statistics from population trajectories}
\usage{
stat_abundance_trend(x, classes = NULL, zscale = TRUE, fn = c(median, sd), ...)

stat_abundance_moment(
  x,
  classes = NULL,
  zscale = TRUE,
  fn = c(median, sd),
  ...
)
}
\arguments{
\item{x}{an object returned from
\code{\link[aae.pop:simulate]{aae.pop::simulate()}} containing
simulated population trajectories}

\item{classes}{a list of integer vectors specifying which how classes
should be grouped to calculate abundances. Groupings do not need to
be disjoint, that is, classes can be included in multiple groups.
Defaults to \code{list(seq_len(nclass), c(2:nclass), c(1))}, which
returns summed abundances of all individuals, all individuals except
the first class, and the first class only}

\item{zscale}{logical determining whether summary statistics are z-scaled.
Defaults to \code{TRUE}}

\item{fn}{function or vector of functions defining how abundances are
summarised over all replicates. Defaults to \code{c(median, sd)},
which calculates the median and standard deviation of abundances
over replicate trajectories}

\item{\dots}{additional arguments passed to fn}
}
\description{
Functions to define summary statistics based on
  observed or simulated population dynamics, which can be used to
  support inference on population dynamics models using
  approximate Bayesian computation.
}
\details{
\code{stat_abundance_trend} returns the abundances of different
  groupings of population classes in each year, averaged over replicate
  trajectories according to \code{fn}. The default is to calculate the
  median and standard deviation of abundances over all replicates, with
  groupings representing all individuals, all individuals except the first
  class (often young-of-year recruits), and the first class only.

\code{stat_abundance_moment} returns the moments over all years
  of population trajectories. Populations are grouped according to
  \code{classes}, as for \code{stat_abundance_trend}. Moments are
  averaged over all replicates according to \code{fn}, which defaults
  to the median and standard deviation. The first four moments are
  included (mean, variance, skewness, kurtosis).
}
\examples{
# to add
# to add
}
